refactor(utils): better management for utils modules to avoid kitchen sink issue
