+++ src/utils/url.ts
@@ -24,7 +24,6 @@ export const getPattern = (label: string): string[] | null => {
 }
 
 export const getPathFromURL = (url: string): string => {
-  // XXX
   const match = url.match(URL_REGEXP)
   if (match) {
     return match[5]
@@ -38,43 +37,4 @@ export const isAbsoluteURL = (url: string): boolean => {
     return true
   }
   return false
-}
-
-const bufferEqual = (a: ArrayBuffer, b: ArrayBuffer) => {
-  if (a === b) {
-    return true
-  }
-  if (a.byteLength !== b.byteLength) {
-    return false
-  }
-
-  const va = new DataView(a)
-  const vb = new DataView(b)
-
-  let i = va.byteLength
-  while (i--) {
-    if (va.getUint8(i) !== vb.getUint8(i)) {
-      return false
-    }
-  }
-
-  return true
-}
-
-export const timingSafeEqual = async (a: any, b: any) => {
-  const sa = await crypto.subtle.digest(
-    {
-      name: 'SHA-256',
-    },
-    new TextEncoder().encode(String(a))
-  )
-
-  const sb = await crypto.subtle.digest(
-    {
-      name: 'SHA-256',
-    },
-    new TextEncoder().encode(String(b))
-  )
-
-  return bufferEqual(sa, sb) && a === b
 }
